Insert a star into a shape with no overlapping visible
How to insert an image , using Paint,Net, so that the image sits on top of the previous image but contained within the boundary of the first image. See file PaintStarInsertion.png. This shows the star and the T shape. The required final result is the 3rd image at the bottom of the page.